Reading the .txt file containing many paths to FASTA files, building an (H)IBF index for FASTA files (kmers), storing the index, exception handling, small tests running ok

+++ include/configuration.hpp
@@ -10,5 +10,5 @@ struct configuration
 {
     std::filesystem::path file_list_path{};
     std::filesystem::path index_output{"index"};
-    size_t kmer_size{20u};
+    uint8_t kmer_size{20u};
 };
